requirements.txt
requests
python-dotenv
questionary
rich
pandas
openpyxl
